refactor: refactor AuthForm
